requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2

# Data processing
numpy>=2.0.0
pandas>=2.2.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0

# Video & face recognition
opencv-python>=4.10.0.84
face-recognition>=1.3.0
Pillow>=10.0.0

# Database
psycopg2-binary==2.9.9

# Environment configuration
python-dotenv==1.0.0

# Future SaaS dependencies (install when ready)
# Flask-Login==0.6.3
# Flask-Mail==0.9.1
# Flask-Migrate==4.0.5
# Flask-RESTful==0.3.10
# Flask-Bcrypt==1.0.1
# celery==5.3.6
# gunicorn==21.2.0

